requirements.txt
Flask==3.0.0
SQLAlchemy==2.0.23
requests==2.31.0
python-dateutil==2.8.2
openpyxl==3.1.5
